Refactoring

+++ src/main/java/dashboard/Output.java
@@ -415,7 +415,7 @@ public class Output implements ObserverOnConstruction, ObserverOnSolving, Observ
 		m.put(DISTRIBUTION, features.ctrTypes);
 		m.separator(features.tableSizes.size() > 0);
 		m.put(TABLES, features.tableSizes.toString());
-		m.put(N_TUPLES, features.tableSizes.repartition.entrySet().stream().mapToLong(e -> e.getValue() * (Integer) e.getKey()).sum());
+		m.put(N_TUPLES, features.tableSizes.repartition.entrySet().stream().mapToLong(e -> e.getValue() * e.getKey()).sum());
 		int nExtStructures = 0, nSharedExtStructures = 0, nIntStructures = 0, nSharedintStructures = 0, nCftStructures = 0, nSharedCftStructures = 0;
 		for (Constraint c : head.problem.constraints) {
 			if (c instanceof ConstraintExtension)
